requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aniso8601==7.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
charset-normalizer==3.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
decorator==4.4.2
Django==3.2.22
django-channels-graphql-ws==0.9.1
django-cors-headers==4.3.0
django-filter==23.3
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.115.0
google-auth==2.24.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
graphene==2.1.9
graphene-django==2.16.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
hyperlink==21.0.0
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
incremental==22.10.0
moviepy==1.0.3
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.2
oauthlib==3.2.2
Pillow==10.1.0
proglog==0.1.10
promise==2.3
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.3
service-identity==23.1.0
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.66.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.2
zope.interface==6.1
